backend/requirements.txt
fastapi
uvicorn[standard]
firecrawl-py
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph
langsmith
pydantic
python-dotenv
typing-extensions
aiohttp
httpx
python-multipart
